requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.3
django-bootstrap5==23.3
django-multiselectfield==0.1.12
Pillow==10.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
Unidecode==1.3.6
